nano-search-mcp/pyproject.toml
[project]
name = "nano-search-mcp"
version = "0.1.0"
description = "NanoSearch MCP Server - 基于 MCP 协议的搜索服务"
requires-python = ">=3.10"
dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0",
    "uvicorn>=0.30.0",
    "playwright>=1.40.0",
    "beautifulsoup4>=4.12.0",
    "markdownify>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[project.scripts]
nano-search-mcp = "nano_search_mcp.__main__:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "B", "UP", "BLE"]
ignore = ["E501"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["BLE001"]
